requirements.txt

streamlit==1.35.0

langchain-groq==0.1.3
langchain-core==0.1.52

python-dotenv==1.0.1